backend/requirements.txt
# FastAPI and core dependencies - Latest stable versions (2024)
fastapi[standard]==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0

# Database and ORM
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10  # PostgreSQL adapter

# Security and authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# HTTP client for external APIs
httpx==0.28.1

# CORS middleware (FastAPI includes this built-in)
# python-cors is not needed

# Development and testing
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
isort==5.13.2
flake8==7.1.1
faker==33.1.0  # For generating dummy data

# Production
gunicorn==23.0.0